Split hiberante and application ehcaches; inject cachemanger into application caches.
Noticed two sets of rss feeds cached in the heap dump.

This revealed that hiberate has it's own cache manager, which was setup before ours.
This forced the snapshots to be stored in a tmp folder, where that are not persisted to disk, which explains the snapshots not persisting issue?