Remove the need for just about everything to push the latest news items onto the model for the right hand slot include.
Came from thinking about how to let the tag template pull in a publisher + tag combiner of a specific page only - want to be able to edit the RHS from the vm files without having to change the java to put stuff on the model.

Moves us closed to the ModelBuilder only having to populate the core content regardless of the display mode (html / rss etc).