Leveraging existing place name tags to increase the number of newsitems which appear in location based searches.
If a resource has no explict geotag of it's own, but is tagged with a tag which has a location attached to it, quietly apply that tag's location to the newsitem.

Refinement: make this happen on a new invisible index field rather than the pin point location field. Newsitems included in the search results by this method don't really have an exactly known point like explictly tagged items do; you'll just end up with an unaccurate cluster on the suburbs center point.