Suggested feeds are sorted by the date of the most recent item available in Whakaoko.
They were been sorted by the Whakaoko.latestItemDate field which is the date of the recent item date on the remote RSS feed;
not the date of that item's persisted copy in Whakaoko.
It makes sense to sort my feed.lastItemDate which is populated from the date of the most recent persisted Whakaoko feed item seen by the feed reader.